Changes in alpha release 0.1.7 (2-Oct-2006) * Water portals rendering works even with the lightmaps. The materials are bad though - not looking as they should. * Added/restructuralised the materials * Added OpdeService, OpdeServiceFactory and some more. These will hopefully be a base classes for the services. * Lightmaps are partially fixed. Swiched on by default now. * Pre-add of the StructInfo sources (Thanks telliamed!). These should form the binary service later on (for universal binary file access using binary templates) * Some sky box work

+++ src/services/worldrep/LightmapAtlas.cpp
@@ -366,7 +366,7 @@ namespace Opde {
 	
 	void LightMap::AddSwitchableLightmap(int id, lmpixel *data) {
 		switchable_lmap.insert(std::make_pair(id, data));
-		intensities.insert(std::make_pair(id, float(0)));
+		intensities.insert(std::make_pair(id, float(1)));
 		
 		// register us
 		owner->registerAnimLight(id, this);
